refactor(dashboard): extract project rename route
Keep the project route module below its file-size ratchet and lower the baseline after extraction.
