+++ scripts/file-size-baseline.txt
@@ -8,7 +8,7 @@
 1530 src/dashboard/frontend/src/components/chat/ConversationPanel.tsx
 1440 src/dashboard/server/routes/command-deck.ts
 1005 src/dashboard/server/routes/flywheel.ts
-1050 src/dashboard/server/routes/projects.ts
+1047 src/dashboard/server/routes/projects.ts
 1143 src/dashboard/server/routes/settings.ts
 1727 src/dashboard/server/routes/workspaces.ts
 1935 src/dashboard/server/routes/workspaces/merge-ops.ts
