+++ src/dashboard/server/routes/projects.ts
@@ -1,10 +1,5 @@
 import { jsonResponse } from "../http-helpers.js";
-/**
- * Projects route module — Effect HttpRouter.Layer (PAN-821)
- *
- * Implements:
- *   GET /api/projects/:projectKey/session-tree
- */
+/** Projects route module — Effect HttpRouter.Layer (PAN-821). */
 
 import { access, readFile, readdir, mkdir, stat, realpath } from 'node:fs/promises';
 import { join, isAbsolute, sep, resolve, normalize, dirname, relative } from 'node:path';
@@ -14,7 +9,8 @@ import { Effect, Layer } from 'effect';
 import { HttpRouter, HttpServerRequest } from 'effect/unstable/http';
 
 import { httpHandler } from './http-handler.js';
-import { resolveProjectFromIssueSync, listProjectsSync, getProjectSync, renameProjectSync, setProjectAutoMergeDefaultSync, setProjectSwarmPolicySync } from '../../../lib/projects.js';
+import { postProjectRenameRoute } from './project-rename.js';
+import { resolveProjectFromIssueSync, listProjectsSync, getProjectSync, setProjectAutoMergeDefaultSync, setProjectSwarmPolicySync } from '../../../lib/projects.js';
 import { resolveSwarmPolicy } from '../../../lib/swarm-policy.js';
 import type { SwarmPolicyLayer } from '../../../lib/swarm-policy.js';
 import { readIssueRecordSync } from '../../../lib/pan-dir/record.js';
@@ -766,46 +762,6 @@ const postProjectAutoMergeDefaultRoute = HttpRouter.add(
     }
     setProjectAutoMergeDefaultSync(key, v);
     return jsonResponse({ value: v });
-  })),
-);
-
-// ─── Route: POST /api/projects/:projectKey/rename ─────────────────────────────
-const postProjectRenameRoute = HttpRouter.add(
-  'POST',
-  '/api/projects/:projectKey/rename',
-  httpHandler(Effect.gen(function* () {
-    const request = yield* HttpServerRequest.HttpServerRequest;
-    const authError = rejectUnsafeDashboardMutationRequest(request);
-    if (authError) return authError;
-
-    const projectKey = (yield* HttpRouter.params)['projectKey'] ?? '';
-    const projects = listProjectsSync();
-    const project = projects.find(({ key }) => key === projectKey)
-      ?? projects.find(({ config }) => config.name === projectKey);
-    if (!project) return jsonResponse({ error: 'Project not found' }, { status: 404 });
-
-    const body = (yield* readProjectJsonBody) as { name?: unknown };
-    if (typeof body.name !== 'string') {
-      return jsonResponse({ error: 'Project name must be a string' }, { status: 400 });
-    }
-
-    try {
-      renameProjectSync(project.key, body.name);
-    } catch (err) {
-      const message = err instanceof Error ? err.message : String(err);
-      if (message.startsWith('Unknown project:')) {
-        return jsonResponse({ error: message }, { status: 404 });
-      }
-      if (message === 'Project name must not be empty') {
-        return jsonResponse({ error: message }, { status: 400 });
-      }
-      if (message.includes('conflicts with existing project')) {
-        return jsonResponse({ error: message }, { status: 409 });
-      }
-      throw err;
-    }
-
-    return jsonResponse({ key: project.key, name: body.name.trim() });
   })),
 );
